Fix unreachable code by removing it or commenting it out
I'd prefer to just remove all unreachable code but some have expressed the desire
to keep the example code around.  Therefore, where appropriate, I commented it
out for now since it silences the cop and doesn't remove the example code.

Note, I believe we can whitelist each of these using a comment:
`# rubocop:disable Lint/UnreachableCode`

But, that implies that the current code is correct and we're just saying "ignore"
it.  Clearly, unreachable code is wrong so let's not pretend this previously
active but unreachable code is correct.

rubocop --only "Lint/UnreachableCode"